--- a/go.mod
+++ b/go.mod
@@ -3,14 +3,10 @@ module github.com/libp2p/go-libp2p-noise
 go 1.12
 
 require (
-	github.com/ChainSafe/log15 v1.0.0
 	github.com/flynn/noise v0.0.0-20180327030543-2492fe189ae6
 	github.com/gogo/protobuf v1.3.1
-	github.com/golang/protobuf v1.3.4
-	github.com/ipfs/go-log v0.0.1
+	github.com/golang/protobuf v1.3.4 // indirect
 	github.com/libp2p/go-libp2p v0.7.2
-	github.com/libp2p/go-libp2p-core v0.3.1
-	github.com/libp2p/go-msgio v0.0.4
+	github.com/libp2p/go-libp2p-core v0.5.0
 	github.com/multiformats/go-multiaddr v0.2.1
-	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550
 )