## Improvements

* make `Read()` conform to `io.Reader` behaviour: `Read()` used to behave like `io.ReadFull()`; it now behaves like `io.Reader` mandates.

* Heavy performance optimisation (see benchmarks and benchcmp below):
  * Preallocate 2-byte (long) slices for message length calculation.
  * on read, copy directly to supplied buffer, if the message is smaller or equal to len(buffer) (zero-alloc path).
  * use buffer pools (via go-buffer-pool) to contain allocs and GC.

* Fix handshake i/o operations not setting deadlines.

* Remove PDFs from repo.

* Wrap errors.

* Refine and improve comments.

## Benchmarks (master vs. this branch)

```
⟩ git checkout 606df358faeaca7509f6556ae5917d8b8ba1f4d1
Note: checking out '606df358faeaca7509f6556ae5917d8b8ba1f4d1'.
HEAD is now at 606df35 bench: reset timer + report allocs.
⟩ go test -count=1 -run=NONE -benchtime=5s -bench  . | tee old.txt
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-noise
BenchmarkTransfer1MB-8          2088       2620798 ns/op     400559109 bytes/sec     4762127 B/op        281 allocs/op
BenchmarkTransfer100MB-8          25     228606052 ns/op     458687650 bytes/sec    472017481 B/op     25746 allocs/op
BenchmarkTransfer500Mb-8           5    1137357804 ns/op     460971180 bytes/sec    2359911987 B/op   128630 allocs/op
BenchmarkHandshakeXX-8          6932        853004 ns/op       23063 B/op        304 allocs/op
PASS
ok      github.com/libp2p/go-libp2p-noise   31.346s

⟩ git checkout raul-review
Previous HEAD position was 606df35 bench: reset timer + report allocs.
Switched to branch 'raul-review'
⟩ go test -count=1 -run=NONE -benchtime=5s -bench  . | tee new.txt
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-noise
BenchmarkTransfer1MB-8          2918       2055000 ns/op     510708724 bytes/sec       36068 B/op         83 allocs/op
BenchmarkTransfer100MB-8          30     199768509 ns/op     524905157 bytes/sec      137920 B/op       6421 allocs/op
BenchmarkTransfer500Mb-8           6     998917048 ns/op     524858261 bytes/sec      567322 B/op      32028 allocs/op
BenchmarkHandshakeXX-8          6806        852974 ns/op       23334 B/op        302 allocs/op
PASS
ok      github.com/libp2p/go-libp2p-noise   37.214s

⟩ benchcmp old.txt new.txt
benchmark                    old ns/op      new ns/op     delta
BenchmarkTransfer1MB-8       2620798        2055000       -21.59%
BenchmarkTransfer100MB-8     228606052      199768509     -12.61%
BenchmarkTransfer500Mb-8     1137357804     998917048     -12.17%
BenchmarkHandshakeXX-8       853004         852974        -0.00%

benchmark                    old allocs     new allocs     delta
BenchmarkTransfer1MB-8       281            83             -70.46%
BenchmarkTransfer100MB-8     25746          6421           -75.06%
BenchmarkTransfer500Mb-8     128630         32028          -75.10%
BenchmarkHandshakeXX-8       304            302            -0.66%

benchmark                    old bytes      new bytes     delta
BenchmarkTransfer1MB-8       4762127        36068         -99.24%
BenchmarkTransfer100MB-8     472017481      137920        -99.97%
BenchmarkTransfer500Mb-8     2359911987     567322        -99.98%
BenchmarkHandshakeXX-8       23063          23334         +1.18%
```

---

Fixes #57. 
Fixes #75.
Fixes #76.