@raulk 2020-04-24
@Stebalien all done here. Final benchmarks vs. benchmarks before your review:

```
⟩ go test -count=1 -run=NONE -benchtime=5s -bench  . | tee new2.txt
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-noise
BenchmarkTransfer1MB-8     	    2832	   2105028 ns/op	 498591263 bytes/sec	   35826 B/op	      83 allocs/op
BenchmarkTransfer100MB-8   	      27	 205335938 ns/op	 510672492 bytes/sec	  140965 B/op	    6421 allocs/op
BenchmarkTransfer500Mb-8   	       5	1033737823 ns/op	 507178476 bytes/sec	  550246 B/op	   32023 allocs/op
BenchmarkHandshakeXX-8     	    7046	    855459 ns/op	   23289 B/op	     300 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-noise	31.242s

⟩ benchcmp new.txt new2.txt
benchmark                    old ns/op     new ns/op      delta
BenchmarkTransfer1MB-8       2055000       2105028        +2.43%
BenchmarkTransfer100MB-8     199768509     205335938      +2.79%
BenchmarkTransfer500Mb-8     998917048     1033737823     +3.49%
BenchmarkHandshakeXX-8       852974        855459         +0.29%

benchmark                    old allocs     new allocs     delta
BenchmarkTransfer1MB-8       83             83             +0.00%
BenchmarkTransfer100MB-8     6421           6421           +0.00%
BenchmarkTransfer500Mb-8     32028          32023          -0.02%
BenchmarkHandshakeXX-8       302            300            -0.66%

benchmark                    old bytes     new bytes     delta
BenchmarkTransfer1MB-8       36068         35826         -0.67%
BenchmarkTransfer100MB-8     137920        140965        +2.21%
BenchmarkTransfer500Mb-8     567322        550246        -3.01%
BenchmarkHandshakeXX-8       23334         23289         -0.19%
```

Not a lot has changed (maybe the compiler was already optimising for us?). Throughput is down a little (maybe an artefact). bytes up and down -- not a lot of difference. Effects may be more noticeable in real-world scenarios? Or have I done something wrong?