@Stebalien 2020-04-24
Note: if the receive buffer is large enough, we can avoid allocating entirely on ready by:

1. Reading into the buffer passed in read.
2. Decrypting in-place.

But it may be best to punt that to a new PR.