@raulk 2020-04-24
Merging this. There are definitely a few more optimisations that we can pursue, but they'll require further refactoring, which I don't have time to do now.

Final benchmarks -- looks like my previous run had some artefacts, this is much better -- compared to master:

```
⟩ go test -count=1 -run=NONE -benchtime=5s -bench  . | tee new3.txt
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-noise
BenchmarkTransfer1MB-8     	    2952	   2015012 ns/op	 520935205 bytes/sec	   35994 B/op	      83 allocs/op
BenchmarkTransfer100MB-8   	      30	 196408553 ns/op	 533885318 bytes/sec	  144692 B/op	    6421 allocs/op
BenchmarkTransfer500Mb-8   	       6	 986187290 ns/op	 531633178 bytes/sec	  563954 B/op	   32027 allocs/op
BenchmarkHandshakeXX-8     	    6944	    864960 ns/op	   23238 B/op	     298 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-noise	28.178s
⟩ benchcmp old.txt new3.txt
benchmark                    old ns/op      new ns/op     delta
BenchmarkTransfer1MB-8       2620798        2015012       -23.11%
BenchmarkTransfer100MB-8     228606052      196408553     -14.08%
BenchmarkTransfer500Mb-8     1137357804     986187290     -13.29%
BenchmarkHandshakeXX-8       853004         864960        +1.40%

benchmark                    old allocs     new allocs     delta
BenchmarkTransfer1MB-8       281            83             -70.46%
BenchmarkTransfer100MB-8     25746          6421           -75.06%
BenchmarkTransfer500Mb-8     128630         32027          -75.10%
BenchmarkHandshakeXX-8       304            298            -1.97%

benchmark                    old bytes      new bytes     delta
BenchmarkTransfer1MB-8       4762127        35994         -99.24%
BenchmarkTransfer100MB-8     472017481      144692        -99.97%
BenchmarkTransfer500Mb-8     2359911987     563954        -99.98%
BenchmarkHandshakeXX-8       23063          23238         +0.76%
```