@raulk 2020-04-24
@Stebalien follow-ups captured here: https://github.com/libp2p/go-libp2p-noise/issues/77.